Add embedded scroll handling (need testing with iOs)
Add handling for "wheel" event with mouse input : disabling zoom on map if CTRL key is not pressed while scrolling, and passing the event to the webpage embedding the map (so that the webpage embedding the map continue its scroll).
Same kind of behaviour with touch event (for mobile) : disabling one finger gestures in favor of two fingers gesture, disabling two fingers map rotation to ease pan/zoom with two fingers.
This behaviors are to avoid scrolling down the webpage embedding the map and accidently zooming the map instead of continuing scrolling up or down (same behaviour as embedded GoogleMaps).

Known issue (to test) : on iOs (last version), with codepen.io on either Chrome or Safari, the webpage jumps back to top when touch events are being handled by the map on my device.